+++ internal/store/mailbox.go
@@ -142,6 +142,9 @@ func initMailboxBucket(tx *bolt.Tx, bucketName []byte) error {
 	if _, err := bucket.CreateBucketIfNotExists(apiIDsBucket); err != nil {
 		return err
 	}
+	if _, err := bucket.CreateBucketIfNotExists(deletedIDsBucket); err != nil {
+		return err
+	}
 
 	return nil
 }
@@ -240,13 +243,7 @@ func (storeMailbox *Mailbox) txGetAPIIDsBucket(tx *bolt.Tx) *bolt.Bucket {
 
 // txGetDeletedIDsBucket returns the bucket with messagesID marked as deleted
 func (storeMailbox *Mailbox) txGetDeletedIDsBucket(tx *bolt.Tx) *bolt.Bucket {
-	// There should be no error since it _...returns an error if the bucket
-	// name is blank, or if the bucket name is too long._
-	bucket, err := storeMailbox.txGetBucket(tx).CreateBucketIfNotExists(deletedIDsBucket)
-	if err != nil || bucket == nil {
-		storeMailbox.log.WithError(err).Error("Cannot create or get bucket with deleted IDs.")
-	}
-	return bucket
+	return storeMailbox.txGetBucket(tx).Bucket(deletedIDsBucket)
 }
 
 // txGetBucket returns the bucket of mailbox containing mapping buckets.